# internal/controller/response.go
package controller

import (
	"github.com/gin-gonic/gin"
	"github.com/sirupsen/logrus"
)

type Error struct {
	Message string `json:"message"`
}

func NewErrorResponse(ctx *gin.Context, statusCode int, message string) {
	logrus.Error(message)

	ctx.AbortWithStatusJSON(statusCode, Error{message})
}